+++ Quicksilver/Code-App/QSPlugInsPrefPane.h
@@ -14,7 +14,7 @@
 	NSMutableArray *plugInArray;
 	NSMutableArray *plugins;
 	NSMutableSet *disabledPlugIns;
-	int viewMode;
+	NSInteger viewMode;
 
 	NSString *search;
 	NSString *category;
@@ -22,8 +22,8 @@
 	IBOutlet NSView *sidebar;
 }
 + (void)getMorePlugIns;
-- (int) viewMode;
-- (void)setViewMode:(int)newViewMode;
+- (NSInteger) viewMode;
+- (void)setViewMode:(NSInteger)newViewMode;
 
 
 - (IBAction)showPlugInsFolder:(id)sender;
@@ -53,7 +53,7 @@
 - (void)reloadFilters;
 - (NSArray *)selectedPlugIns;
 
-- (void)setViewMode:(int)newViewMode;
+- (void)setViewMode:(NSInteger)newViewMode;
 - (void)reloadFiltersIgnoringViewMode:(BOOL)ignoreView;
 
 - (id)preferencesSplitView;
