+++ Quicksilver/Code-App/QSTriggersPrefPane.h
@@ -49,7 +49,7 @@
 
 	QSTrigger *selectedTrigger;
 	NSSplitView *splitView;
-	int selectedRow;
+	NSInteger selectedRow;
 }
 - (NSString *)currentSet;
 - (void)setCurrentSet:(NSString *)value;
@@ -78,8 +78,8 @@
 
 - (IBAction)showTriggerInfo:(id)sender;
 - (IBAction)hideTriggerInfo:(id)sender;
-- (int)tabViewIndex;
-- (void)setTabViewIndex:(int)index;
+- (NSInteger)tabViewIndex;
+- (void)setTabViewIndex:(NSInteger)index;
 
 - (IBAction)editCommand:(id)sender;
